Upload in first commit - two examples of files (transmission and reception) in arduino, using SoftwareSerial
Two arduino Uno boards, use the SoftwareSerial library and are connected serially (TTL serial - direct connection). The idea is to promote the on-off control of a LED, through the logic state at the input of the other Uno board. In order for the switching of logic levels to occur on the control board, a push-button switch is installed on a pin configured as pull-up.
A very short video with the demonstration can be found at this author's link: https://youtu.be/C70NpzlpS3U